+++ R/helper_test_MGD.R
@@ -757,3 +757,306 @@ getRelevantParameters <- function(
 
 
 
+#' Internal: get structured cSEMTestMGD results
+#'
+#' Convenience function to summarize the results of all tests resulting from a 
+#' call to [testMGD()] in a user-friendly way.
+#' 
+#' @usage structureTestMGDDecisions(.object)
+#'
+#' @inheritParams csem_arguments
+#' 
+#' @return A data.frame.
+#' 
+#' @keywords internal
+
+
+
+structureTestMGDDecisions <- function(.object){
+  
+  ## Install dplyr, tidyr, and rlang if not already installed
+  if (!requireNamespace("dplyr", quietly = TRUE)) {
+    stop2(
+      "Package `dplyr` required. Use `install.packages(\"dplyr\")` and rerun.")
+  }
+  if (!requireNamespace("tidyr", quietly = TRUE)) {
+    stop2(
+      "Package `tidyr` (> 1.0.0) required. Use `install.packages(\"tidyr\")` and rerun.")
+  }
+  if (!requireNamespace("rlang", quietly = TRUE)) {
+    stop2(
+      "Package `rlang` required. Use `install.packages(\"rlang\")` and rerun.")
+  }
+  
+  # Check if of class cSEMTestMGD
+  if(!inherits(.object, "cSEMTestMGD")){
+    stop2("The following error occured in the `structureTestMGDDecisions()` function:\n",
+          "Object must be of class `cSEMResults`")
+  }
+  
+  # Helper function ------------------------------------------------------------
+  
+  # Change CI alpha equivlanet (1-CI)
+  changeCItoAlphaEquivalent <- function(.names){
+    # .names <- c("99%","95%","90%")
+    tmpNames <- as.numeric(gsub("%", "", .names))
+    tmpNames <- 100-tmpNames
+    tmpNames <- paste0(tmpNames, "%")
+    return(tmpNames)
+  }
+  
+  out <- dplyr::tibble()
+  # Klesel ---------------------------------------------------------------------
+  
+  # Test only provides an overall decision 
+  
+  if("Klesel" %in% names(.object)){
+    klesel <- .object$Klesel$Decision %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "alpha") %>%
+      # longer format
+      tidyr::pivot_longer(cols = c(.data$dG, .data$dL), 
+                          names_to = "Distance_metric", 
+                          values_to = "Decision") %>%
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Klesel", Comparison = "overall")
+    
+    out <- dplyr::bind_rows(out, klesel)
+  }
+  
+  # Chin  ----------------------------------------------------------------------
+  
+  if("Chin" %in% names(.object)){
+    
+    # overall decision
+    chin1 <- .object$Chin$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::mutate(Test = "Chin", Comparison = "overall")
+    
+    chin2 <-  .object$Chin$Decision %>% 
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "Comparison") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "alpha") %>%
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      # Check all comparisons
+      dplyr::group_by(.data$`p-value_correction`, .data$alpha) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison), all) %>%
+      dplyr::ungroup() %>% 
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$`p-value_correction`, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Chin")
+    
+    out <- dplyr::bind_rows(out, chin1, chin2)
+  }
+  
+  # Sarstedt  ------------------------------------------------------------------
+  
+  if("Sarstedt" %in% names(.object)){
+    
+    # overall result
+    sarstedt1 <- .object$Sarstedt$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::mutate(Test = "Sarstedt", Comparison = "overall")
+    
+    # decision based on single paths
+    sarstedt2 <- .object$Sarstedt$Decision %>%
+      purrr::modify_depth(2, dplyr::bind_rows) %>%
+      purrr::map(dplyr::bind_rows, .id = "alpha") %>%
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$`p-value_correction`, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Sarstedt")
+    
+    out <- dplyr::bind_rows(out, sarstedt1, sarstedt2)
+  }
+  
+  # Keil  ----------------------------------------------------------------------
+  
+  if("Keil" %in% names(.object)){
+    
+    # overall decision
+    keil1 <- .object$Keil$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::mutate(Test = "Keil", Comparison = "overall")
+    
+    # decision based on single paths
+    keil2 <- .object$Keil$Decision %>%
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "Comparison") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "alpha") %>%
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::group_by(.data$`p-value_correction`, .data$alpha) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison),all) %>%
+      dplyr::ungroup() %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$`p-value_correction`, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Keil")
+    
+    out <- dplyr::bind_rows(out, keil1, keil2)
+  }
+  
+  # Nitzl  ---------------------------------------------------------------------
+  
+  if("Nitzl" %in% names(.object)){
+    
+    # overall decision
+    nitzl1 <- .object$Nitzl$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::mutate(Test = "Nitzl", Comparison = "overall")
+    
+    # decision based on single paths
+    nitzl2 <- .object$Nitzl$Decision %>%
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "Comparison") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "alpha") %>%
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::group_by(.data$`p-value_correction`, .data$alpha) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison),all) %>%
+      dplyr::ungroup() %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$`p-value_correction`, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Nitzl")
+    
+    out <- dplyr::bind_rows(out, nitzl1, nitzl2)
+  }
+  
+  # Henseler  ------------------------------------------------------------------
+  
+  if("Henseler" %in% names(.object)){
+    
+    # overall decision
+    henseler1 <- .object$Henseler$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::mutate(Test = "Henseler", Comparison = "overall")
+    
+    # decision based on single paths
+    henseler2 <- .object$Henseler$Decision %>%
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "Comparison") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "alpha") %>%
+      dplyr::bind_rows(.id = "p-value_correction") %>%
+      dplyr::group_by(.data$`p-value_correction`, .data$alpha) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison),all) %>%
+      dplyr::ungroup() %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$`p-value_correction`, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      dplyr::mutate(Test = "Henseler")
+    
+    out <- dplyr::bind_rows(out, henseler1, henseler2)
+  }
+  
+  # CI para  -------------------------------------------------------------------
+  
+  if("CI_para" %in% names(.object)){
+    
+    # overall decision
+    CIpara1 <- .object$CI_para$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "alpha") %>%
+      tidyr::pivot_longer(cols = dplyr::contains("CI"), 
+                          names_to = "CI_type", 
+                          values_to = "Decision") %>%
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      # Change % as equivlanet to alpha
+      dplyr::rename_at(dplyr::vars(dplyr::contains("%")), changeCItoAlphaEquivalent) %>%
+      dplyr::mutate(Test = "CI_para", Comparison = "overall")
+    
+    # decision based on single paths
+    CIpara2 <-  .object$CI_para$Decision %>%
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(3, ~ dplyr::select(., .data$Name, .data$Decision)) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "CI_type") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "Comparison") %>%
+      dplyr::bind_rows(.id = "alpha") %>%
+      tidyr::pivot_wider(names_from = .data$Name, values_from = .data$Decision) %>%
+      dplyr::group_by(.data$alpha, .data$CI_type) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison), all) %>%
+      dplyr::ungroup() %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$CI_type, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      # Change % as equivalent to alpha
+      dplyr::rename_at(dplyr::vars(dplyr::contains("%")), changeCItoAlphaEquivalent) %>%
+      dplyr::mutate(Test = "CI_para")
+    
+    out <- dplyr::bind_rows(out, CIpara1, CIpara2)
+    
+  }
+  # CI overlap  ----------------------------------------------------------------
+  
+  if("CI_overlap" %in% names(.object)){
+    
+    # overall decision
+    CIoverlap1 <- .object$CI_overlap$Decision_overall %>% 
+      purrr::map(dplyr::bind_rows) %>% 
+      dplyr::bind_rows(.id = "alpha") %>%
+      tidyr::pivot_longer(cols = dplyr::contains("CI"), 
+                          names_to = "CI_type", 
+                          values_to = "Decision") %>%
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      # Change % as equivalent to alpha
+      dplyr::rename_at(dplyr::vars(dplyr::contains("%")), changeCItoAlphaEquivalent) %>%
+      dplyr::mutate(Test = "CI_overlap", Comparison = "overall")
+    
+    # decision based on single paths
+    CIoverlap2 <-  .object$CI_overlap$Decision %>%
+      purrr::modify_depth(3, dplyr::bind_rows) %>%
+      purrr::modify_depth(3, ~ dplyr::select(., .data$Name, .data$Decision)) %>%
+      purrr::modify_depth(2, dplyr::bind_rows, .id = "CI_type") %>%
+      purrr::modify_depth(1, dplyr::bind_rows, .id = "Comparison") %>%
+      dplyr::bind_rows(.id = "alpha") %>%
+      tidyr::pivot_wider(names_from = .data$Name, values_from = .data$Decision) %>%
+      dplyr::group_by(.data$alpha, .data$CI_type) %>%
+      dplyr::summarise_at(dplyr::vars(-.data$Comparison), all) %>%
+      dplyr::ungroup() %>%
+      # put results into format 
+      tidyr::pivot_longer(cols = (-c(.data$CI_type, .data$alpha)),
+                          names_to = "Comparison", 
+                          values_to = "Decision") %>%
+      # put alphas in cols
+      tidyr::pivot_wider(names_from = .data$alpha, values_from = .data$Decision) %>%
+      # Change % as equivalent to alpha
+      dplyr::rename_at(dplyr::vars(dplyr::contains("%")), changeCItoAlphaEquivalent) %>%
+      dplyr::mutate(Test = "CI_overlap")
+    
+    out <- dplyr::bind_rows(out, CIoverlap1, CIoverlap2)
+  }
+  
+  # Summarize all results ------------------------------------------------------
+  
+  out <- dplyr::select(out, .data$Test, .data$Comparison, dplyr::contains("%"), 
+                       .data$`p-value_correction`, .data$CI_type, 
+                       .data$Distance_metric)
+  
+  return(out)
+}
+
+